Updating return models
