--- docs/_docs/Oauth2/Get Current Application Information.md
@@ -1,24 +1,0 @@
----
-title: Get Current Application Information
-category: Oauth2
-order: 1
----
-
-# `getCurrentApplicationInformation`
-
-```php
-$client->oauth2->getCurrentApplicationInformation($parameters);
-```
-
-## Description
-
-
-
-## Parameters
-
-No Parameters
-
-## Response
-
-Returns the bot&#039;s OAuth2 application info.
-
