+++ docs/_docs/Webhook/Execute Webhook.md
@@ -12,6 +12,7 @@ $client->webhook->executeWebhook($parameters);
 
 ## Description
 
+warn
 This endpoint supports both JSON and form data bodies. It does require multipart/form-data requests instead of the normal JSON request type when uploading files. Make sure you set your Content-Type to multipart/form-data if you&#039;re doing that. Note that in that case, the embeds field cannot be used, but you can pass an url-encoded JSON body as a form value for payload_json.
 
 ## Parameters
